[ File Roller 2.23.5 released ]
2008-08-04  Paolo Bacchilega  <paobac@svn.gnome.org>

	[ File Roller 2.23.5 released ]

	* NEWS: updated for 2.23.5

2008-08-04  Paolo Bacchilega  <paobac@svn.gnome.org>

	* src/file-utils.c: return TRUE if path is NULL

svn path=/trunk/; revision=2398

+++ src/file-utils.c
@@ -875,6 +875,9 @@ remove_local_directory (const char *path)
 	char     *uri;
 	gboolean  result;
 
+	if (path == NULL)
+		return TRUE;
+	
 	uri = g_filename_to_uri (path, NULL, NULL);
 	result = remove_directory (uri);
 	g_free (uri);
